% Generated by roxygen2 (4.1.1): do not edit by hand
\name{.DollarNames.Octave}
\alias{.DollarNames,Octave-method}
\alias{.DollarNames.Octave}
\title{Auto-completion for Octave Interface Object}
\usage{
\S3method{.DollarNames}{Octave}(x, pattern = "")

\S4method{.DollarNames}{Octave}(x, pattern = "")
}
\arguments{
  \item{x}{ An R object for which valid names after \code{"$"}
    are computed and returned.
  }

  \item{pattern}{ A regular expression.  Only matching names are
    returned.
  }
}
\description{
  Auto-completion for \code{\linkS4class{Octave}} objects
}
\keyword{internal}

