% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{.isPlatformCompatible}
\alias{.isPlatformCompatible}
\title{Platform Compatibility Check for RcppOctave}
\usage{
.isPlatformCompatible(details = FALSE)
}
\arguments{
\item{details}{logical that indicates if the checks' details should be 
returned as well, in which case, the result is a list.}
}
\description{
Checks if the current platform supports RcppOctave.
}
\details{
The following checks are performed:
\itemize{
\item platform must be either non-Windows or Windows-i386;  
\item The Octave version to be loaded is the same one as the one against
which RcppOctave libraries and modules were compiled.
}
}
