% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-functions.R
\name{o_config_info}
\alias{o_config_info}
\title{Embedded Octave Configuration and Installation Information}
\usage{
o_config_info(var = c("CC", "CC_VERSION", "FC"))
}
\arguments{
\item{var}{name of the configuration variable to retrieve.}
}
\description{
Retrieves configuration and installation information about 
the embedded instance of Octave currently used by \pkg{RcppOctave}.
}
\note{
In Octave 4.0.0., variables \code{CC_VERSION} and \code{CXX_VERSION} 
were renamed \code{GCC_VERSION} and \code{GXX_VERSION} respectively.
\code{o_config_info} supports both values, querying the right ones
depending on the linked Octave version.  
 
See \url{http://www.gnu.org/software/octave/NEWS-4.0.html}
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}


o_config_info()
# NB: some config variables are not available on all versions
try( o_config_info('USE_64_BIT_IDX_T') )

}
\seealso{
Other Octave.info: \code{\link{Octave.config}},
  \code{\link{Octave.home}}, \code{\link{octave_config}}
}
