% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Octave-class.R
\name{o_exist}
\alias{o_exist}
\title{Checking Octave Variables}
\usage{
o_exist(NAME, ...)
}
\arguments{
\item{NAME}{name to check existence.}

\item{...}{extra parameters passed to the Octave function 
\code{exist}.}
}
\description{
Checks if an Octave object of a given name exists,
using the Octave function \code{exist}.
}
\section{Octave Documentation for \emph{exist}}{

\Sexpr[results=rd,stage=render]{if( .Platform$OS.type != 'windows' || .Platform$r_arch != 'x64' ) RcppOctave::o_help(exist, format='rd')}

\emph{[Generated from Octave-\Sexpr{RcppOctave::o_version()} on \Sexpr{Sys.time()}]}
}

