% Generated by roxygen2 (4.1.1): do not edit by hand
\docType{class}
\name{Octave-class}
\alias{Octave-class}
\title{Class Octave: Seamless Access to Octave Functions and Variables}
\description{
  This class provides a direct interface to Octave functions and base context
(i.e. the scope where user objects are defined).
}
\keyword{internal}

