% Generated by roxygen2 (4.1.1): do not edit by hand
\docType{data}
\name{Octave.config}
\alias{Octave.config}
\alias{Octave.info}
\alias{Octave.version}
\title{Octave Configuration and Installation Information}
\format{\preformatted{List of 15
 $ bindir        : chr "/usr/bin"
 $ platform      : chr "x86_64-pc-linux-gnu"
 $ version       : chr "3.8.1"
 $ api_version   : chr "api-v49+"
 $ cc            : chr "gcc"
 $ f77           : chr "gfortran"
 $ modules       : chr "/home/renaud/Documents/projects/RcppOctave/libtest/RcppOctave/modules"
 $ lflags        : chr "-L/usr/lib/x86_64-linux-gnu/octave/3.8.1 -L/usr/lib/x86_64-linux-gnu"
 $ .oct_ldflags  : chr " -L/usr/lib/x86_64-linux-gnu/octave/3.8.1 -L/usr/lib/x86_64-linux-gnu -loctinterp -loctave "
 $ .oct_cppflags : chr "-D_FORTIFY_SOURCE=2 -I/usr/include/octave-3.8.1/octave/.. -I/usr/include/octave-3.8.1/octave -DOCT_POST_3_4_0=1"
 $ libs          : chr [1:2] "liboctinterp" "liboctave"
 $ libdir        : chr [1:2] "/usr/lib/x86_64-linux-gnu/octave/3.8.1" "/usr/lib/x86_64-linux-gnu"
 $ customed      : logi FALSE
 $ version.string: chr "Octave version 3.8.1 (api-v49+)"
 $ home          : chr "/usr"
 - attr(*, "class")= chr "simple.list"
}}
\usage{
Octave.config

Octave.version

Octave.info(name)
}
\arguments{
  \item{name}{name of the variable to retrieve}
}
\description{
  The functions documented here enable retrieving information
about the Octave installation used at installation or runtime
-- which should normally be the same.

\code{Octave.info} is a function that retrieves information about
the version of Octave that is used by the current session
of \pkg{RcppOctave}.
}
\details{
  \code{Octave.config} is a list that extends \code{Octave.version} with
extra information about compilers and compilation flags.

\code{Octave.version} is list that contains version information as determined
by the configure script at installation time.
}
\examples{
\dontshow{# roxygen generated flag
options(R_CHECK_RUNNING_EXAMPLES_=TRUE)
}

Octave.config

Octave.version

Octave.info()
Octave.info('modules')
}
\seealso{
  Other Octave.info: \code{\link{Octave.home}};
  \code{\link{o_config_info}}; \code{\link{octave_config}}
}
\keyword{datasets}

