% Generated by roxygen2 (4.1.1): do not edit by hand
\docType{package}
\name{RcppOctave}
\alias{RcppOctave}
\alias{RcppOctave-package}
\title{Seamless Interface to Octave -- And Matlab}
\description{
  Direct interface to Octave. The primary goal is to facilitate the
port of Matlab/Octave scripts to R. The package enables to call any Octave
functions from R and as well as browsing their documentation, passing
variables between R and Octave, using R core RNGs in Octave, which ensures
that stochastic computations are also reproducible.

The primary goal is to facilitate the port of Matlab/Octave scripts to R.
The package enables to call any Octave functions from R and as well as browsing
their documentation, passing variables between R and Octave, using R core RNGs
in Octave, which ensure stochastic computations are also reproducible.
}
\details{
  \tabular{ll}{
Package: \tab RcppOctave\cr
Type: \tab Package\cr
Version: \tab 0.18.3\cr
Date: \tab 2015-11-10\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\examples{
\dontshow{# roxygen generated flag
options(R_CHECK_RUNNING_EXAMPLES_=TRUE)
}

.CallOctave('svd', matrix(1:9, 3))
o_help('svd')
}
\author{
  Renaud Gaujoux

Maintainer: Renaud Gaujoux <renaud@tx.technion.ac.il>

Renaud Gaujoux \email{renaud@tx.technion.ac.il}

Maintainer: Renaud Gaujoux \email{renaud@tx.technion.ac.il}
}
\seealso{
  See \code{\link{.CallOctave}}, \code{\link{o_source}}, \code{\link{o_help}}
}
\keyword{package}

