% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{rcppoctave}
\alias{rcppoctave}
\title{inline C++ Function with Octave Integration}
\usage{
rcppoctave(..., plugin = "RcppOctave")
}
\arguments{
\item{...}{arguments passed to \code{\link[inline]{rcpp}}}
}
\description{
inline C++ Function with Octave Integration
}
\examples{


# call from R
set.seed(1)
res1 <- .O$rand(10)

# call from c++
feval <- RcppOctave:::rcppoctave(signature(), '
Rcpp::List args = Rcpp::List::create(10);
  SEXP result = CallOctave("rand", args);
  return(result);
')

set.seed(1)
res2 <- feval()
stopifnot( identical(res1, res2) ) 

}
