\docType{methods}
\name{show,Octave-method}
\alias{show,Octave-method}
\title{Show method for \code{\linkS4class{Octave}} objects}
\usage{
  \S4method{show}{Octave}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
  Show method for \code{\linkS4class{Octave}} objects
}
\keyword{methods}

