% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Octave-class.R
\docType{class}
\name{Octave-class}
\alias{Octave-class}
\title{Class Octave: Seamless Access to Octave Functions and Variables}
\description{
This class provides a direct interface to Octave functions and base context
(i.e. the scope where user objects are defined).
}
\keyword{internal}
