\name{Octave}
\alias{Octave}
\alias{Octave-class}
\alias{show}
\alias{show,OctaveFunction-method}
\alias{show,Octave-method}
\title{Class Octave: Seamless Access to Octave Functions and Variables}
\description{
  This class provides a direct interface to Octave
  functions and base context (i.e. the scope where user
  objects are defined).
}
\keyword{internal}

