% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{Octave.config}
\alias{Octave.config}
\alias{Octave.version}
\alias{Octave.info}
\title{Octave Configuration and Installation Information}
\format{An object of class \code{simple.list} of length 15.}
\usage{
Octave.config

Octave.version

Octave.info(name)
}
\arguments{
\item{name}{name of the variable to retrieve}
}
\description{
The functions documented here enable retrieving information 
about the Octave installation used at installation or runtime
-- which should normally be the same.

\code{Octave.info} is a function that retrieves information about 
the version of Octave that is used by the current session 
of \pkg{RcppOctave}.
}
\details{
\code{Octave.config} is a list that extends \code{Octave.version} with 
extra information about compilers and compilation flags.

\code{Octave.version} is list that contains version information as determined
by the configure script at installation time.
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}

Octave.config
Octave.version
Octave.info()
Octave.info('modules')

}
\seealso{
Other Octave.info: \code{\link{Octave.home}},
  \code{\link{o_config_info}}, \code{\link{octave_config}}
}
\keyword{datasets}
