% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{sourceExamples}
\alias{sourceExamples}
\title{Loading Example M-files}
\usage{
sourceExamples(file)
}
\arguments{
\item{file}{filename of the example script to source. If missing, the function 
lists all the m-files from the \dQuote{scripts/} sub-directory.}
}
\description{
Source an example M-file in the sub-directory \dQuote{scripts/} of RcppOctave
installation.
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}


sourceExamples()
sourceExamples('ex_source.m')

}
